+++ src/F1Telemetry.cpp
@@ -1,8 +1,10 @@
 #include "F1Telemetry.h"
 #include "AboutDialog.h"
 #include "CheckUpdatesDialog.h"
+#include "F1TelemetrySettings.h"
 #include "FileDownloader.h"
-#include "SettingsKeys.h"
+#include "PreferencesDialog.h"
+#include "ThemeDialog.h"
 #include "Tracker.h"
 #include "ui_F1Telemetry.h"
 
@@ -10,19 +12,22 @@
 #include <QMessageBox>
 #include <QNetworkReply>
 #include <QPushButton>
+#include <QScrollBar>
 #include <QTextEdit>
 #include <QTimer>
 
 F1Telemetry::F1Telemetry(QWidget *parent) : QMainWindow(parent), ui(new Ui::F1Telemetry)
 {
+	initDefaultSettings();
+
 	ui->setupUi(this);
+
+	updateTheme();
 
 	setWindowIcon(QIcon(":/Ressources/F1Telemetry.png"));
 	setWindowTitle(qApp->applicationName() + " " + qApp->applicationVersion());
 
 	_tracker = new Tracker(this);
-
-	initDefaultSettings();
 
 	buildListener();
 
@@ -54,18 +59,18 @@ F1Telemetry::~F1Telemetry() { delete ui; }
 void F1Telemetry::buildListener()
 {
 	delete _listener;
-	auto address = QSettings().value(SERVER).toString();
-	auto port = QSettings().value(PORT).toInt();
-	_listener = new F1Listener(_tracker, address, port, this);
+	F1TelemetrySettings settings;
+	_listener = new F1Listener(_tracker, settings.server(), settings.port(), this);
 	ui->trackingWidget->setConnectionStatus(_listener->isConnected());
 }
 
 void F1Telemetry::loadSettings()
 {
-	QSettings settings;
+	F1TelemetrySettings settings;
 	ui->trackingWidget->loadSettings(&settings);
 	ui->compareLapsWidget->loadSettings(&settings);
 	ui->compareStintsWidget->loadSettings(&settings);
+	ui->compareRaceWidget->loadSettings(&settings);
 
 	restoreGeometry(settings.value("windowGeometry").toByteArray());
 	restoreState(settings.value("windowState").toByteArray());
@@ -74,10 +79,11 @@ void F1Telemetry::loadSettings()
 
 void F1Telemetry::saveSetings()
 {
-	QSettings settings;
+	F1TelemetrySettings settings;
 	ui->trackingWidget->saveSettings(&settings);
 	ui->compareLapsWidget->saveSettings(&settings);
-	ui->compareStintsWidget->loadSettings(&settings);
+	ui->compareStintsWidget->saveSettings(&settings);
+	ui->compareRaceWidget->saveSettings(&settings);
 
 	settings.setValue("windowGeometry", saveGeometry());
 	settings.setValue("windowState", saveState());
@@ -86,15 +92,16 @@ void F1Telemetry::saveSetings()
 
 void F1Telemetry::initDefaultSettings()
 {
-	QSettings settings;
-	if(!settings.contains(PORT))
-		settings.setValue(PORT, 20777);
-	if(!settings.contains(SERVER))
-		settings.setValue(SERVER, "");
+	F1TelemetrySettings settings;
+	settings.init();
 }
 
 void F1Telemetry::initMenu()
 {
+	auto prefMenu = ui->menuBar->addMenu("&Preferences");
+	prefMenu->addAction("Charts Theme...", this, &F1Telemetry::editTheme);
+	prefMenu->addAction("User Preferences...", this, &F1Telemetry::editPreferences);
+
 	auto helpMenu = ui->menuBar->addMenu("&Help");
 
 	helpMenu->addAction("About &Qt", [=]() { QMessageBox::aboutQt(this, qApp->applicationName()); });
@@ -120,6 +127,8 @@ bool F1Telemetry::isGreaterVersion(const QString &version)
 	for(int i = 0; i < nbElement; ++i) {
 		if(newVersion.value(i, "0").toInt() > oldVersion.value(i, "0").toInt()) {
 			return true;
+		} else if(newVersion.value(i, "0").toInt() < oldVersion.value(i, "0").toInt()) {
+			return false;
 		}
 	}
 
@@ -132,19 +141,20 @@ void F1Telemetry::closeEvent(QCloseEvent *event)
 	QMainWindow::closeEvent(event);
 }
 
-void F1Telemetry::startTracking(bool trackPlayer,
-								bool trackTeammate,
-								bool trackAllCars,
-								const QVector<int> &trackedDriverIds)
+void F1Telemetry::startTracking(const TrackedCars &cars)
 {
 	_tracker->clearTrackedDrivers();
-	if(trackPlayer)
+	if(cars.me)
 		_tracker->trackPlayer();
-	if(trackTeammate)
+	if(cars.player2)
+		_tracker->trackPlayer2();
+	if(cars.teammate)
 		_tracker->trackTeammate();
-	if(trackAllCars)
+	if(cars.all)
 		_tracker->trackAllCars();
-	for(auto id : trackedDriverIds)
+	if(cars.allRace)
+		_tracker->trackAllRace();
+	for(auto id : cars.ids)
 		_tracker->trackDriver(id);
 	_tracker->setDataDirectory(ui->trackingWidget->getDataDirectory());
 	_tracker->start();
@@ -163,7 +173,7 @@ void F1Telemetry::fileDownloaded(int type, const QByteArray &data)
 		if(isGreaterVersion(data)) {
 
 			qInfo() << "A newer version is available";
-			QSettings settings;
+			F1TelemetrySettings settings;
 			if(!_isAutoCheckUpdates || settings.value("skipedVersion") != data) {
 				_downloader->downloadFile(
 					QUrl("https://bitbucket.org/Fiingon/pxg-f1-telemetry/raw/master/Changelog.md"), ChangelogFile);
@@ -211,6 +221,7 @@ void F1Telemetry::showChangeLog()
 		simplifyRegexp.setMinimal(true);
 		html.remove(simplifyRegexp);
 		edit->document()->setHtml(html);
+		edit->moveCursor(QTextCursor::Start);
 
 		dialog.resize(700, 700);
 		dialog.exec();
@@ -219,9 +230,8 @@ void F1Telemetry::showChangeLog()
 
 void F1Telemetry::changelogAutoDisplay()
 {
-	QSettings settings;
-	if(!settings.allKeys().isEmpty() &&
-	   settings.value("lastChangelogAutoDisplay").toString() != qApp->applicationVersion()) {
+	F1TelemetrySettings settings;
+	if(!settings.isEmpty() && settings.value("lastChangelogAutoDisplay").toString() != qApp->applicationVersion()) {
 		QTimer::singleShot(0, this, &F1Telemetry::showChangeLog);
 		settings.setValue("lastChangelogAutoDisplay", qApp->applicationVersion());
 	}
@@ -243,4 +253,34 @@ void F1Telemetry::contact()
 	layout->addItem(horizontalSpacer, layout->rowCount(), 0, 1, layout->columnCount());
 
 	msg.exec();
+}
+
+void F1Telemetry::editTheme()
+{
+	ThemeDialog dialog(this);
+	if(dialog.exec() == QDialog::Accepted) {
+		updateTheme();
+	}
+}
+
+void F1Telemetry::updateTheme()
+{
+	F1TelemetrySettings settings;
+	if(settings.useCustomTheme()) {
+		auto customTheme = settings.customTheme();
+		ui->compareLapsWidget->setCustomTheme(customTheme);
+		ui->compareStintsWidget->setCustomTheme(customTheme);
+		ui->compareRaceWidget->setCustomTheme(customTheme);
+	} else {
+		auto theme = settings.theme();
+		ui->compareLapsWidget->setTheme(theme);
+		ui->compareStintsWidget->setTheme(theme);
+		ui->compareRaceWidget->setTheme(theme);
+	}
+}
+
+void F1Telemetry::editPreferences()
+{
+	PreferencesDialog dialog(this);
+	dialog.exec();
 }