+++ src/TelemetryDataTableModel.cpp
@@ -8,21 +8,7 @@
 
 #include <QtDebug>
 
-const QList<QColor> BASIC_COLOR_LIST = {QColor(41, 141, 198), QColor(153, 202, 83), QColor(246, 166, 83), QColor(190, 95, 213)};
-
-TelemetryDataTableModel::TelemetryDataTableModel()
-{
-	for(int i = 0; i < 5; ++i) {
-		for(const auto &color : BASIC_COLOR_LIST) {
-			_availableColors << color.darker(100 + i * 20);
-		}
-	}
-	for(int i = 0; i < 5; ++i) {
-		for(const auto &color : BASIC_COLOR_LIST) {
-			_availableColors << color.darker(100 - i * 20);
-		}
-	}
-}
+TelemetryDataTableModel::TelemetryDataTableModel(const QList<QColor> &baseColors) { setBaseColors(baseColors); }
 
 QVariant TelemetryDataTableModel::data(const QModelIndex &index, int role) const
 {
@@ -159,4 +145,39 @@ void TelemetryDataTableModel::exportRowToCsv(int index, QString path)
     QtConcurrent::run([data, path] {
         data->exportData(path);
     });
+}
+
+void TelemetryDataTableModel::setVisibleAllExcept(int row, bool value)
+{
+	for(auto i = 0; i < _visibility.count(); ++i) {
+		if(value) {
+			_visibility[i] = i != row;
+		} else {
+			_visibility[i] = i == row;
+		}
+	}
+
+	emit dataChanged(QModelIndex(), QModelIndex());
+	emit visibilityChanged();
+}
+
+void TelemetryDataTableModel::setVisibleAll(bool value) { setVisibleAllExcept(-1, value); }
+
+void TelemetryDataTableModel::setBaseColors(const QList<QColor> &colors)
+{
+	_availableColors.clear();
+	for(int i = 0; i < 5; ++i) {
+		for(const auto &color : colors) {
+			_availableColors << color.darker(100 + i * 10);
+		}
+	}
+	for(int i = 1; i < 5; ++i) {
+		for(const auto &color : colors) {
+			_availableColors << color.darker(100 - i * 10);
+		}
+	}
+
+	for(int i = 0; i < _colors.count(); ++i) {
+		_colors[i] = _availableColors.takeFirst();
+	}
 }